systemd service: enable systemd-remount-fs in containers
Disabling this service in containers precludes setting custom options
for e.g., /proc.  For context, see discussion at
https://github.com/NixOS/nixpkgs/pull/14372.

All the service does is remount / and API file systems using options
found in /etc/fstab (if any are set), so there shouldn't be any
downsides to running this within a container.