tor-browser: rewrite
This rewrite brings an improved wrapper, more precise patching of
run-time dependencies, and better support for pluggable transports.
Because we end up using very little of the standard build phases, the
builder is now a stand-alone script.  The wrapper has become more
complicated and is also factored out, to avoid having to deal with
quoting issues etc.

Some specific highlights:
- The wrapper writes all its state to $TBB_HOME, which defaults to
  $XDG_DATA_HOME/tor-browser
  - The user's state is versioned, avoiding the problem where the user
  continues to use old extension data after an update.
- The wrapper makes more of an effort to run in a cleaned up environment
- The application and extension auto-updaters have been disabled
- Use specific paths to torrc and friends to avoid permission problems
  and allow tbb to update its torrc
- PaX support

Currently, obfs* and fte are the only supported pluggable transports;
meek and scramblesuit do not work.