urbit: support bootstrap
urbit can use a pre-compiled Arvo binary to bootstrap itself.  Unfortunately,
urbit expects the binary to be in the current working directory, where it also
will write state.  To work around this, generate a wrapper that cds into
URBIT_HOME, copying the pre-compiled arvo binary into it if necessary, before
calling out to the real urbit.  This seems to work for me at least.